Lane generator changes (#42)
* Added Mesh, Renamed MeshFactory, Added MeshDrawComponent and added interface called DrawComponent

* GTFO Texture!

* Added base lane generating

* Added documentation

* Rewrote LaneGenerator to work with queue

* Before the wave of merge conflicts

* Removed merge error

* Changes?

* Added moving LaneObstacles

* Added moving LaneObstacles 2.0

* Fixed float problem in loading lengths

+++ ProftaakPeriode4/ProftaakPeriode4/MeshFactory.cpp
@@ -235,14 +235,14 @@ Mesh* LoadMeshFile(const std::string& fileName)
 				LoadMaterialFile(dirName + "/" + params[1], dirName, indexedMesh);
 			}else if(params[0] == "w")
 			{
-				indexedMesh->_width = float(std::stoi(params[1]));
+				indexedMesh->_width = std::stof(params[1]);
 			}else if(params[0] == "l")
 			{
-				indexedMesh->_length = float(std::stoi(params[1]));
+				indexedMesh->_length = std::stof(params[1]);
 			}
 			else if (params[0] == "h")
 			{
-				indexedMesh->_height = float(std::stoi(params[1]));
+				indexedMesh->_height = std::stof(params[1]);
 			}
 			else if (params[0] == "usemtl")
 			{
